testing trailing slash, made edits to styling
changed how the intersection observer works for it to work with various screen sizes autumatically rather than having to hard code magical values for each threshold it crosses. also made changes to how content is being shown in the home page with styleHome, having it resize based on viewport width or show content in the middle of the page. almost finished, just need to add mores styling for various screen sizes and will be done.